+++ src/pages/index.astro
@@ -11,7 +11,7 @@ const { data } = await supabase.from("covers").select();
 >
   <div class="main">
     <h1>Genderswap.fm</h1>
-    <p>Work in progress.</p>
+    <p>Some covers deliver the age-old simple pleasures of drag…</p>
     {data?.map((cover) => <a href={`cover/${cover.slug}`}>{cover.slug}</a>)}
   </div>
 </BaseLayout>
@@ -32,4 +32,3 @@ const { data } = await supabase.from("covers").select();
     }
   }
 </style>
- ../lib/supabase
